requirements.txt
asgiref==3.4.1
beautifulsoup4==4.9.3
Django==3.2.5
django-bootstrap-v5==1.0.4
django-braces==1.14.0
django-model-utils==4.1.1
djangorestframework==3.12.4
importlib-metadata==3.10.1
Pillow==8.3.1
psycopg2==2.9.1
pytz==2021.1
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
swapper==1.1.2.post1
zipp==3.5.0
